Update rsshelper.sh
There was an issue with the ed script causing it to not populate new entries properly, causing line items to be scrambled around in the wrong places. Honestly not really sure why due to poor ed documentation resources. I reworked the script logic to instead insert new items at the bottom of the rss feed. This appears to work better.